# api/openapi.yaml
openapi: 3.0.3
info:
  title: Agora API
  version: 1.0.0
  description: |
    Agora School Management Platform API.
    Base contract aligned with Step 3 (auth, attendance, homework, marks, messaging, notifications).
servers:
  - url: /api/v1
    description: Relative base URL
...
paths:
...
  /notifications:
    get:
      tags: [Notifications]
      operationId: listNotifications
      summary: List notifications for current user
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/NotificationStatus'
        - name: channel
          in: query
          schema:
            $ref: '#/components/schemas/NotificationChannel'
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/pageSizeQuery'
      responses:
        '200':
          description: Notification list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /notifications/{notificationId}/read:
    patch:
      tags: [Notifications]
      operationId: markNotificationRead
      summary: Mark one notification as read
      parameters:
        - $ref: '#/components/parameters/notificationIdPath'
      responses:
        '200':
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /notifications/test:
    post:
      tags: [Notifications]
      operationId: sendTestNotification
      summary: Send test notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestNotificationRequest'
      responses:
        '200':
          description: Notification queued/sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /notifications/push-tokens:
    get:
      tags: [Notifications]
      operationId: listPushTokens
      summary: List push device tokens for current user
      responses:
        '200':
          description: Push tokens list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokenListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Notifications]
      operationId: upsertPushToken
      summary: Register or refresh push device token for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PushTokenUpsertRequest'
      responses:
        '200':
          description: Push token upserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PushTokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /notifications/push-tokens/{tokenId}:
    delete:
      tags: [Notifications]
      operationId: deletePushToken
      summary: Revoke one push token for current user
      parameters:
        - name: tokenId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Push token revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooleanResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /internal/notifications/trigger:
    post:
      tags: [Internal]
      operationId: internalTriggerNotification
      summary: Internal event trigger for notifications
      security:
        - InternalApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalNotificationTriggerRequest'
      responses:
        '200':
          description: Internal trigger accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooleanResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceApiKey:
      type: apiKey
      in: header
      name: X-Device-Api-Key
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
  parameters:
    pageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSizeQuery:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    notificationIdPath:
      name: notificationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
    PagingMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    ErrorItem:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
...
    NotificationChannel:
      type: string
      enum: [in_app, push, email, sms]
    NotificationStatus:
      type: string
      enum: [queued, sent, failed, read]
    PushProvider:
      type: string
      enum: [fcm]
    PushPlatform:
      type: string
      enum: [android, ios, web]
...
    SuccessBooleanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            ok:
              type: boolean
              example: true
        meta:
          $ref: '#/components/schemas/Meta'
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        channel:
          $ref: '#/components/schemas/NotificationChannel'
        status:
          $ref: '#/components/schemas/NotificationStatus'
        payload:
          type: object
          additionalProperties: true
        sent_at:
          type: string
          format: date-time
          nullable: true
        read_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    NotificationResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Notification'
        meta:
          $ref: '#/components/schemas/Meta'
    NotificationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Notification'
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/PagingMeta'
    TestNotificationRequest:
      type: object
      required: [user_id, title, body, channel]
      properties:
        user_id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        channel:
          $ref: '#/components/schemas/NotificationChannel'
    PushToken:
      type: object
      properties:
        id:
          type: string
          format: uuid
        provider:
          $ref: '#/components/schemas/PushProvider'
        platform:
          $ref: '#/components/schemas/PushPlatform'
        device_id:
          type: string
          nullable: true
        app_version:
          type: string
          nullable: true
        is_active:
          type: boolean
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PushTokenUpsertRequest:
      type: object
      required: [provider, platform, device_token]
      properties:
        provider:
          $ref: '#/components/schemas/PushProvider'
        platform:
          $ref: '#/components/schemas/PushPlatform'
        device_token:
          type: string
        device_id:
          type: string
        app_version:
          type: string
    PushTokenResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/PushToken'
        meta:
          $ref: '#/components/schemas/Meta'
    PushTokenListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/PushToken'
        meta:
          $ref: '#/components/schemas/Meta'
...
    InternalNotificationTriggerRequest:
      type: object
      required: [event_type, school_id, actor_user_id, payload]
      properties:
        event_type:
          type: string
          example: attendance.marked
        school_id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
security:
  - BearerAuth: []